protection for nil-pointers
In github/github.go, around LOC128, some fields from `user` are used:
  user.Login, user.Name, user.URL
These fields are *string, and it was seen that it might happen that in
some cases they may have "nil" as value, in which cases the intent to
de-reference them (*user.Name) will cause a panic in the go-code-execution.
To account for these situations, I've added a function that will return
an empty string if the pointer-to-string is nil.

Signed-off-by: Paulo Campos <paulo.aleixodacosta@bestsecret.com>